+++ tests/warnings/test_warnings.py
@@ -37,6 +37,7 @@ def get_files_from_folder(foldername):
 #
 #    assert errors.num_messages()!=0
 
+@pytest.mark.xdist_incompatible
 @pytest.mark.parametrize("f",get_files_from_folder('semantic'))
 def test_semantic_warnings(f):